Add files via upload
Cipher programs are for encryption and decryption of a string using caesar cipher. The ServerCipher program takes the String from a user and encrypts the string using caesar cipher and sends the encrypted string to the connected client. The ClientCipher program takes the encrypted string from the server and decrypts the string with socketprogramming.